refactor(app): apis, added kopcuj.sql
