+++ frontend/src/pages/admin.js
@@ -88,7 +88,7 @@ function AdminPage() {
 
     //Fetch reviews from database
     const fetchReviews = async () => {
-        const response = await axios.get(`http://localhost:8082/api/review/`);
+        const response = await axios.get(`http://localhost:8082/api/reviews/`);
         return response.data;
     }
