+++ frontend/src/pages/map.js
@@ -108,7 +108,7 @@ function MapPage() {
     }
 
     const fetchReviews = async () => {
-        const response = await axios.get(`http://localhost:8082/api/review/`);
+        const response = await axios.get(`http://localhost:8082/api/reviews/`);
         return response.data;
     }
